requirements.txt
Django==5.1.6
django-allauth==65.14.0
django-redis==5.4.0
django-simple-captcha==0.6.2
psycopg2-binary==2.9.10
gunicorn==23.0.0
Pillow==11.1.0
redis==5.2.1
requests==2.32.3
django-simpleui==2025.1.13
django-crispy-forms==2.3
crispy-bootstrap4==2024.10
celery==5.6.0
django-celery-beat==2.7.0
django-celery-results==2.6.0
django-haystack==3.3.0
Whoosh==2.7.4
jieba==0.42.1
djangorestframework==3.14.0
django-imagekit==5.0.0
markdown-checklist==0.4.4
pypinyin==0.53.0  
user_agent==0.1.10
docker==7.1.0
bleach==6.2.0
beautifulsoup4==4.13.3
pymdown-extensions==10.14.3
pycryptodome==3.21.0
flower==2.0.1
gevent==25.9.1
psutil==6.1.1
django-storages==1.14.4
boto3==1.35.90
tinycss2==1.5.1
kubernetes>=28.1.0
openai>=1.0.0
anthropic>=0.18.0       
zhipuai>=2.0.0         
qianfan>=0.3.0         
dashscope>=1.14.0
openpyxl==3.1.5
packaging==25.0
